projects/01-image-classification/pyproject.toml
[project]
name = "01-image-classification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi>=0.118.3",
    "numpy>=2.3.3",
    "pillow>=11.3.0",
    "prometheus-client>=0.23.1",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.11.0",
    "python-multipart>=0.0.20",
    "tritonclient[http]>=2.60.0",
    "uvicorn>=0.37.0",
]

[dependency-groups]
lint = [
    "isort>=6.1.0",
    "ruff>=0.14.0",
]
export = [
    "onnx>=1.19.1",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]

# --- BUILD --- #

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
start = "app.main:run"

[tool.hatch.build]
include = ["app/**/*.py", "static/**"]

# --- Tools --- #
[tool.uv]
required-environments = [
    "sys_platform == 'linux' and platform_machine == 'x86_64'",
]
[tool.uv.workspace]
members = ["01-image-classification"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D", "INP001", "PGH004", "EXE002"]
